Fix FA2 fp16 dtype crash and harden CARL live harness
The from-scratch Triton flash-attention kernel crashed in fp16 (the GPU
serving dtype) with "Both operands must be same dtype. Got fp32 and fp16":
tl.dot accumulates in fp32, so the score/probability tensor p is fp32 while
v is loaded fp16, and the second tl.dot(p, v) rejects the mismatch. Cast v
(and the additive mask bias) up to fp32 at the use site; both are no-ops in
fp32 mode, so the atol=1e-4 parity contract is preserved.

This is exactly what made the CARL live cell (docs/run_benchmarks.ipynb 6c)
produce no results on a Colab GPU: --live runs real TinyLlama on CUDA, which
routes attention through this kernel in fp16 and crashed mid-run.

Also harden run_live so it can never fail silently: flush stdout as it goes
(buffered output is lost if the process is OOM-killed), guard each scenario
so a failure prints its full traceback and the surviving scenarios still
render a clean table, and only emit delta lines for scenarios that ran.

+++ src/engine/kernels/flash_attention.py
@@ -196,14 +196,20 @@ def flash_attention_kernel(
                 mask=(offs_m[:, None] < S_q) & (offs_n[None, :] < S_k),
                 other=0.0,
             )
-            s = s + bias
+            # s is fp32 (tl.dot accumulator); cast the bias up so the add stays
+            # fp32 even when the mask tensor is fp16. No-op in fp32 mode.
+            s = s + bias.to(tl.float32)
 
         # --- online softmax update ---
         m_new = tl.maximum(m_i, tl.max(s, axis=1))      # new running max
         p = tl.exp(s - m_new[:, None])                  # this block's weights
         alpha = tl.exp(m_i - m_new)                     # correction for the past
         l_i = l_i * alpha + tl.sum(p, axis=1)
-        acc = acc * alpha[:, None] + tl.dot(p, v, allow_tf32=False)
+        # tl.dot requires both operands to share a dtype. In fp16 mode v is
+        # loaded as fp16 while p is fp32 (tl.dot above accumulates in fp32, and
+        # the online-softmax exp keeps it fp32), so cast v up to fp32 here. In
+        # fp32 mode this is a no-op, so parity is unaffected.
+        acc = acc * alpha[:, None] + tl.dot(p, v.to(tl.float32), allow_tf32=False)
         m_i = m_new
 
     # Normalize. No real query row is ever fully masked (every row attends to
